# backend/doc/openapi.yaml
openapi: 3.0.1
info:
  title: Stupro 2019 Querimonia API
  description: Querimonia ist eine Plattform zum automatischen Beschwerdemanagement. Über die hier beschriebene API kann mit dem Backend kommuniziert werden. Dies beinhaltet Upload von neuen Beschwerden, Abrufen der gespeicherten Beschwerden, Anfragen von Antworttexten, etc.
  version: 0.5.2
servers:
  - url: https://beschwerdemanagement.iao.fraunhofer.de/dev/
  - url: https://beschwerdemanagement.iao.fraunhofer.de/
  - url: https://querimonia.iao.fraunhofer.de/dev
  - url: https://querimonia.iao.fraunhofer.de/
  - url: http://localhost:8080/
...
paths:
...
  /api/actions:
    get:
      tags:
        - actions
      summary: Gibt die gespeicherten Aktionen zurück.
      description: Gibt die gespeicherten Aktionen zurück. Es kann gefiletert und sortiert, sowie Pagination angewandt werden.
      parameters:
        - $ref: '#/components/parameters/CountParameter'
        - $ref: '#/components/parameters/PageParameter'
        - name: sort_by
          in: query
          description: Hier können alle Kategorien angegeben werden, nach denen die Anfrage sortiert werden soll. Die Reihenfolge im Array bestimmt die Priorität beim Sortieren.
          required: false
          schema:
            type: array
            items:
              type: string
              enum: [id_desc, id_asc, name_desc, name_asc]
        - $ref: '#/components/parameters/ActionCodeParameter'
        - $ref: '#/components/parameters/KeywordsParameter'

      operationId: getActions
      responses:
        200:
          description: Erfolgreiche Anfrage.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Action'
        400:
          description: Üngültige Angabe von Parametern.

    post:
      tags:
        - actions
      summary: Fügt dem System eine neue Aktion hinzu.
      description: Fügt eine neue Aktion hinzu.
      requestBody:
        description: Die neue Aktion, die dem System hinzufügt werden soll. Im Request-Body muss die ID nicht angegeben werden, diese wird bei Angabe nicht beachtet.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Action'
      operationId: addAction
      responses:
        201:
          description: Das Aktion ist gültig und wurde gespeichert. Die gespeicherte Aktion wird zurückgegeben.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Action'
        400:
          description: Der Aktionscode oder die Parameter sind ungültig.
  /api/actions/{actionId}:
    parameters:
      - $ref: '#/components/parameters/ActionIdParameter'

    get:
      tags:
        - actions
      summary: Gibt eine Aktion zurück.
      description: Gibt die Aktion mit der gegebenen ID zurück.
      operationId: getAction
      responses:
        200:
          description: Die Aktion wird zurückgegeben.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Action'
        404:
          description: Die Aktion existiert nicht.

    delete:
      tags:
        - actions
      summary: Löscht eine Aktion.
      description: Löscht die gegebene Aktion aus der Datenbank.
      operationId: deleteAction
      responses:
        204:
          description: Der Aktion wurde gelöscht.
        404:
          description: Der Aktion existiert nicht.

    put:
      tags:
        - actions
      summary: Aktualisiert eine bestehende Aktion.
      description: Aktualisiert die Aktion mit der gegebenen ID.
      requestBody:
        description: Die Aktion, die übernommen werden soll. Die ID sollte nicht angegeben werden und wird ansonsten ignoriert.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Action'
      operationId: updateAction
      responses:
        200:
          description: Konfiguration wurde erfolgreich aktualisiert und wird zurückgegeben.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Action'
        404:
          description: Aktion mit der gegebenen ID ist nicht vorhanden.
        400:
          description: Die übergebene Aktion ist nicht gültig.
  /api/actions/count:
    get:
      tags:
        - actions
      summary: Gibt die Anzahl der gespeicherten Aktionen zurück.
      description: Gibt die Anzahl der gespeicherten Aktionen zurück. Es kann nach Aktion-Code und Schlüsselwörtern gefiltert werden.
      parameters:
        - $ref: '#/components/parameters/ActionCodeParameter'
        - $ref: '#/components/parameters/KeywordsParameter'

      operationId: countActions
      responses:
        200:
          description: Erfolgreiche Anfrage.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Action'
        400:
          description: Üngültige Angabe von Parametern.
...
components:
  parameters:
...
    ActionIdParameter:
      name: actionId
      in: path
      required: true
      description: Die eindeutige ID einer Aktion
      schema:
        type: integer
        format: int32
        minimum: 1
        example: 1
    CountParameter:
      name: count
      in: query
      description: Anzahl der Elemente, die zurück gegeben werden sollen. Falls nicht angegeben, werden alle Elemente zurückgegeben.
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        example: 10
    PageParameter:
      name: page
      in: query
      description: Falls page 0 ist, werden die ersten *count* Elemente zurück gegeben, falls page 1 ist die zweiten *count* Elemente, gibt also die Seitenzahl an. Falls nicht angegeben, wird 0 angenommen. Wird ignoriert, falls *count* nicht gegeben ist.
      required: false
      schema:
        type: integer
        format: int32
        default: 0
        minimum: 0
        example: 0
...
    KeywordsParameter:
      name: keywords
      in: query
      required: false
      description: Bei Angabe werden nur Elemente ausgewertet, die die angebenen Schlüsselwörter enthalten.
      schema:
        type: array
        items:
          type: string
        example:
          - "Busfahrer"
          - "unfreundlich"
...
    ActionCodeParameter:
      name: action_code
      in: query
      required: false
      description: Bei Angabe werden nur Aktionen mit dem gegebenen Aktionscode zurückgegeben.
      schema:
        $ref: '#/components/schemas/ActionCode'
  schemas:
...
    Action:
      description: Eine Aktion kann automatisch beim Abschluss einer Beschwerde ausgelöst werden, wenn bestimmt Regeln dafür erfüllt sind. Zum Beispiel können E-Mails versendet werden, Gutscheine als Attachment angehängt werden, Beschwerden weitergeleitet werden etc.
      type: object
      required:
        - name
        - actionCode
        - rulesXml
        - parameters
      properties:
        actionId:
          description: Die eindeutige ID der Aktion.
          type: integer
          format: int32
          minimum: 1
          example: 1
        name:
          description: Ein eindeutiger Bezeichnert für die Aktion.
          type: string
          example: "GutscheinFreifahrt"
        actionCode:
          $ref: '#/components/schemas/ActionCode'
        rulesXml:
          description: Die Regeln, wann eine Aktion ausgelöst werden kann im XML-Format.
          type: string
          format: xml
          example: '<Rules><Subject value="Fahrt nicht erfolgt"/></Rules>'
        parameters:
          description: Zusätzliche Parameter für die Aktion, zum Beispiel an wen eine E-Mail gesendet wird oder den Wert des Gutscheins.
          type: object
          additionalProperties:
            type: string
          example:
            "E-Mail": "beispiel@gmail.com"
            "Wert": "5"
    ActionCode:
      description: Eine Bezeichung, die den Typ der Aktion angibt.
      type: string
      enum:
        - SEND_MAIL
        - ATTACH_VOUCHER
      example: "ATTACH_VOUCHER"
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic